+++ mongopb/README.md
@@ -1,0 +1,3 @@
+# mongopb
+
+Package mongopb provides protobuf utilities for MongoDB.
