+++ mongopb/examples/.gitignore
