+++ mongopb/mongopb.go
@@ -1,0 +1,3 @@
+// Package mongopb provides protobuf utilities for MongoDB.
+
+package mongopb
