--- pbutilmongo/README.md
@@ -1,3 +1,0 @@
-# pbutilmongo
-
-implement me
