--- pbutilmongo/pbutilmongo.go
@@ -1,3 +1,0 @@
-// Package pbutilmongo provides protobuf utilities for MongoDB.
-
-package pbutilmongo
