+++ proto/goinsane/pbutil/mongo.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 package goinsane.pbutil;
-option go_package = "github.com/goinsane/pbutil/pbutilmongo";
+option go_package = "github.com/goinsane/pbutil/mongopb";
 
 message ObjectID {
   string value = 1;
